% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{cond_cmpl}
\alias{cond_cmpl}
\title{Conditional Compilation}
\usage{
cond_cmpl(ipa)
}
\arguments{
\item{ipa}{String. A sequence of IPA symbols.}
}
\description{
\code{cond_cmpl} wraps a sequence of IPA string
with LaTeX code in R Markdown document when compiled
to LaTeX. When compiled to HTML, returns the
original sequence. Call
\code{rmarkdown::draft("file.Rmd", "support_ipa", package = "linguisticsdown")}
for details.
}
