% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/include_graphics2.R
\name{include_graphics_latex}
\alias{include_graphics_latex}
\title{Wrapper of \code{knitr::include_graphics} for PDF Output}
\usage{
include_graphics_latex(path, alt_path = NULL, handler = function(path)
  knitr::asis_output(paste("View", tools::file_ext(path), "at", path)),
  ...)
}
\description{
Deals with URL and GIFs. If an url is passed to
\code{path} of \code{\link[knitr]{include_graphics}},
the figure is automatically downloaded and included
using local relative path. If a figure with \code{.gif}
extension is included, a piece of text, rather than the
figure, is inserted.
}
\keyword{internal}
