% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shinyGadget-xsampa2ipa.R
\name{writeIPA}
\alias{writeIPA}
\title{Insert IPA symbols with Shiny app}
\source{
\url{https://github.com/dmort27/epitran/blob/master/epitran/data/ipa-xsampa.csv}
}
\usage{
writeIPA()
}
\value{
Inserted plain text at the cursor returned by
  \link[rstudioapi]{insertText}.
}
\description{
`writeIPA` opens a shiny gadget in the viewer pane of RStudio
to let users insert a sequence of IPA symbols into
the source pane or console.
Users can choose to use phonetic features (such as
\emph{aspirated}, \emph{schwa},
\emph{vl}(voiceless) / \emph{vd}(voiced), etc.) to find IPA
symbols or use the
\href{https://en.wikipedia.org/wiki/X-SAMPA}{X-SAMPA}
input method directly. Note that due to the special
meanings of backslash(`\\`) in programming,
\strong{backslashes(`\\`) in X-SAMPA symbols are
replaced with slashes(`/`)}.
}
